+++ src/snowflake/snowpark/files.py
@@ -3,18 +3,23 @@
 # Copyright (c) 2012-2025 Snowflake Computing Inc. All rights reserved.
 #
 
-"""
-SnowflakeFile for UDFs and stored procedures in Snowpark.
-
-This class is intended for usage within stored procedures and UDFs and many methods do not work locally.
-"""
 from __future__ import annotations
 
 import array
-import io
 import sys
 import tempfile
-from io import RawIOBase
+import os
+from io import (
+    RawIOBase,
+    UnsupportedOperation,
+    BufferedReader,
+    SEEK_SET,
+    SEEK_END,
+    SEEK_CUR,
+)
+from snowflake.snowpark._internal.utils import RELATIVE_PATH_PREFIX
+from typing import Sequence
+import logging
 
 # Python 3.8 needs to use typing.Iterable because collections.abc.Iterable is not subscriptable
 # Python 3.9 can use both
@@ -24,7 +29,14 @@ if sys.version_info <= (3, 9):
 else:
     from collections.abc import Iterable
 
-_DEFER_IMPLEMENTATION_ERR_MSG = "SnowflakeFile currently only works in UDF and Stored Procedures. It doesn't work locally yet."
+_NON_LOCAL_PATH_ERR_MSG = "SnowflakeFile currently supports only relative paths and read apis in local testing mode."
+_WRITE_MODE_ERR_MSG = (
+    "SnowflakeFile currently doesn't support write APIs in local testing mode."
+)
+_DEFER_IMPLEMENTATION_ERR_MSG = "Not yet supported in UDF and Stored Procedures."
+READ_MODES = ["r", "rb"]
+WRITE_MODES = ["w", "wb"]
+_logger = logging.getLogger(__name__)
 
 
 class SnowflakeFile(RawIOBase):
@@ -35,7 +47,45 @@ class SnowflakeFile(RawIOBase):
 
     The constructor of this class is not supposed to be called directly. Call :meth:`~snowflake.snowpark.file.SnowflakeFile.open` to create a read-only SnowflakeFile object, and call :meth:`~snowflake.snowpark.file.SnowflakeFile.open_new_result` to create a write-only SnowflakeFile object.
 
-    This class is intended for usage within UDFs and stored procedures and many methods do not work locally.
+    This class is used to read and write files in UDFs and stored procedures. On Snowflake, it is used to read and write stage files. It also
+    supports Python IOBase and BufferedBase methods such as :meth:`read`, :meth:`write`, :meth:`close`.
+
+    To read from a staged file, use the following API:
+
+    Example::
+        >>> from snowflake.snowpark.files import SnowflakeFile
+        >>> from snowflake.snowpark.functions import udf
+        >>> @udf
+        ... def read_file(url: str) -> str:
+        ...     file = SnowflakeFile.open(url, "r")
+        ...     return file.read()
+
+    To write to a staged file first write to a result file via the following example.
+    The result file will return as a scoped URL which can be copied to a permanent stage
+    with `copy files <https://docs.snowflake.com/en/sql-reference/sql/copy-files>`_ or read directly via another call to SnowflakeFile.read() in another UDF invocation.
+    See `writing files from Snowpark Python UDFs <https://docs.snowflake.com/en/developer-guide/snowpark/python/creating-udfs#label-snowpark-python-udf-write-files>`_ for more details.
+
+    Example::
+        >>> from snowflake.snowpark.files import SnowflakeFile
+        >>> from snowflake.snowpark.functions import udf
+        >>> @udf
+        ... def write_file(content: str) -> str:
+        ...     file = SnowflakeFile.open_new_result("w")
+        ...     file.write(content)
+        ...     return file # file must be returned to be accessible
+
+    These examples are using the client, but this same pattern can be used inside SQL-defined UDFs.
+
+    We provide a local implementation of SnowflakeFile to aid in local testing. This currently only supports using read APIs on relative paths.
+    Local testing to Snowflake stages is not yet supported.
+
+    Note:
+        1. All of the implementation in this file is for local testing purposes.
+
+        2. There may be slight implementation differences between local testing and Snowflake execution environments, which we call out below when we can.
+        If any issues or these differences block your testing workflow, please file a bug report at https://github.com/snowflakedb/snowpark-python/issues.
+
+        3. UDF implementation is dependent on the Snowflake release. Therefore, this documentation may not always be up to date.
     """
 
     def __init__(
@@ -61,6 +111,22 @@ class SnowflakeFile(RawIOBase):
         self.buffer = None
         self.encoding = None
         self.errors = None
+
+        _DEFAULT_READ_BUFFER_SIZE = 32 * 1024
+        if mode in READ_MODES:
+            # Buffered Reader used to support BufferedIOBase methods such as read1 and readinto1
+            self._file_stream = BufferedReader(
+                open(self._file_location, self._mode), _DEFAULT_READ_BUFFER_SIZE
+            )
+        elif mode in WRITE_MODES:
+            # need to still open a file stream
+            self._file_stream = open(self._file_location, self._mode)
+        self._pos = 0
+        self._is_local_file = (
+            True
+            if self._file_location.startswith((RELATIVE_PATH_PREFIX, "C:"))
+            else False
+        )
 
     @classmethod
     def open(
@@ -89,7 +155,7 @@ class SnowflakeFile(RawIOBase):
             is_owner_file: (Deprecated) A boolean value, if True, the API is intended to access owner's files and all URI/URL are allowed. If False, the API is intended to access files passed into the function by the caller and only scoped URL is allowed.
             require_scoped_url: A boolean value, if True, file_location must be a scoped URL. A scoped URL ensures that the caller cannot access the UDF owners files that the caller does not have access to.
         """
-        if mode not in ("r", "rb"):
+        if mode not in READ_MODES:
             raise ValueError(
                 f"Invalid mode '{mode}' for SnowflakeFile.open. Supported modes are 'r' and 'rb'."
             )
@@ -107,7 +173,7 @@ class SnowflakeFile(RawIOBase):
         Args:
             mode: A string used to mark the type of an IO stream. Supported modes are "w" for text write and "wb" for binary write.
         """
-        if mode not in ("w", "wb"):
+        if mode not in WRITE_MODES:
             raise ValueError(
                 f"Invalid mode '{mode}' for SnowflakeFile.open_new_result. Supported modes are 'w' and 'wb'."
             )
@@ -118,11 +184,48 @@ class SnowflakeFile(RawIOBase):
             from_result_api=True,
         )
 
+    def _raise_if_not_read(self) -> None:
+        """
+        Internal function to validate read mode of the file object before performing an IO operation.
+        """
+        if self._mode not in READ_MODES:
+            raise UnsupportedOperation(f"Not readable mode={self._mode}")
+
+    def _raise_if_not_write(self) -> None:
+        """
+        Internal function to validate write mode of the file object before performing a IO operation.
+        """
+        if self._mode not in WRITE_MODES:
+            raise UnsupportedOperation(f"Not writable mode={self._mode}")
+
+    def _raise_if_closed(self) -> None:
+        """
+        Internal function to validate open status of the file object before performing an IO operation.
+        """
+        if self._file_stream.closed:
+            raise ValueError("I/O operation on closed file.")
+
+    def _read_into_buffer(self, b: bytes | bytearray | array.array) -> int:
+        """
+        Internal function to read bytes into a pre-allocated, writable bytes-like object buffer.
+        This is used by readinto and readinto1 methods.
+        """
+        buffer_len = len(b)
+        if buffer_len == 0:
+            return 0
+        content = self.read1(buffer_len)
+        size = memoryview(content)
+        self._pos += size.nbytes
+        b[:buffer_len] = content
+        return size.nbytes
+
     def close(self) -> None:
         """
-        In UDF and Stored Procedures, the close func closes the IO Stream included in the SnowflakeFile.
+        See https://docs.python.org/3/library/io.html#io.IOBase.close
+
+        Closes the underlying IO Stream of the SnowflakeFile.
         """
-        return
+        self._file_stream.close()
 
     def detach(self) -> None:
         """
@@ -130,112 +233,234 @@ class SnowflakeFile(RawIOBase):
 
         See https://docs.python.org/3/library/io.html#io.BufferedIOBase.detach
         """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        self._raise_if_closed()
+        raise UnsupportedOperation("Detaching stream from file is unsupported")
 
-    def fileno(self) -> None:
+    def fileno(self) -> int:
         """
-        See https://docs.python.org/3/library/io.html#io.IOBase.fileno
+        Getting a file descriptor number is not supported in Snowflake. Raises an OSError.
         """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        self._raise_if_closed()
+        raise OSError("This object does not use a file descriptor")
 
     def flush(self) -> None:
         """
-        Not yet supported in UDF and Stored Procedures.
+        Fail if the stream is closed. Does nothing
         """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        self._raise_if_closed()
+        pass
 
-    def isatty(self) -> None:
+    def isatty(self) -> bool:
         """
         Returns False, file streams in stored procedures and UDFs are never interactive in Snowflake.
         """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        self._raise_if_closed()
+        return False
 
-    def read(self, size: int = -1) -> None:
+    def read(self, size: int = -1) -> Sequence:
         """
-        See https://docs.python.org/3/library/io.html#io.RawIOBase.read
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        From https://docs.python.org/3/library/io.html#io.RawIOBase.read
 
-    def read1(self, size: int = -1) -> None:
-        """
-        See https://docs.python.org/3/library/io.html#io.BufferedIOBase.read1
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        Read up to size bytes from the object and return them. As a convenience, if size is unspecified or -1,
+        all bytes until EOF are returned.
+        Fewer than size bytes may be returned.
 
-    def readable(self) -> None:
+        If 0 bytes are returned, and size was not 0, this indicates end of file.
         """
-        See https://docs.python.org/3/library/io.html#io.IOBase.readable
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        self._raise_if_closed()
+        self._raise_if_not_read()
+        if self._is_local_file:
+            content = self._file_stream.raw.read(size)
+            self._pos += len(content)
+            return content
+        raise NotImplementedError(_NON_LOCAL_PATH_ERR_MSG)
 
-    def readall(self) -> None:
+    def read1(self, size: int = -1) -> Sequence:
         """
-        See https://docs.python.org/3/library/io.html#io.RawIOBase.readall
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        From https://docs.python.org/3/library/io.html#io.BufferedIOBase.read1
 
-    def readinto(self, b: bytes | bytearray | array.array) -> None:
-        """
-        See https://docs.python.org/3/library/io.html#io.IOBase.readinto
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        Read and return up to size bytes, with at most one call to the underlying raw stream’s read() (or readinto()) method.
 
-    def readinto1(self, b: bytes | bytearray | array.array) -> None:
+        If size is -1 (the default), an arbitrary number of bytes are returned (more than zero unless EOF is reached).
         """
-        See https://docs.python.org/3/library/io.html#io.BufferedIOBase.readinto1
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        self._raise_if_closed()
+        self._raise_if_not_read()
+        if self._is_local_file:
+            if self._mode == "r":
+                content = self.read(size).encode()
+            else:
+                content = self._file_stream.read1(size)
+            self._pos += len(content)
+            return content
+        raise NotImplementedError(_NON_LOCAL_PATH_ERR_MSG)
 
-    def readline(self, size: int = -1) -> None:
+    def readable(self) -> bool:
         """
-        See https://docs.python.org/3/library/io.html#io.IOBase.readline
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        From https://docs.python.org/3/library/io.html#io.IOBase.readable
 
-    def readlines(self, hint: int = -1) -> None:
+        Returns whether or not the stream is readable.
         """
-        See https://docs.python.org/3/library/io.html#io.IOBase.readlines
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        self._raise_if_closed()
+        return self._file_stream.readable()
 
-    def seek(self, offset: int, whence: int = io.SEEK_SET) -> int:
+    def readall(self) -> Sequence:
+        """
+        From https://docs.python.org/3/library/io.html#io.RawIOBase.readall
+
+        Read and return all the bytes from the stream until EOF, using multiple calls to the stream if necessary.
+        """
+        return self.read(
+            -1
+        )  # Python IO uses readall as the default implementation to read with no args, so we can just call read(-1)
+
+    def readinto(self, b: bytes | bytearray | array.array) -> int:
+        """
+        From https://docs.python.org/3/library/io.html#io.RawIOBase.readinto
+
+        Read bytes into a pre-allocated, writable bytes-like object b, and return the number of bytes read. For example, b might
+        be a bytearray.
+        """
+        self._raise_if_closed()
+        self._raise_if_not_read()
+        if self._is_local_file:
+            if self._mode == "r":
+                return self._read_into_buffer(b)
+            size = self._file_stream.raw.readinto(b)
+            self._pos += size
+            return size
+        raise NotImplementedError(_NON_LOCAL_PATH_ERR_MSG)
+
+    def readinto1(self, b: bytes | bytearray | array.array) -> int:
+        """
+        From https://docs.python.org/3/library/io.html#io.BufferedIOBase.readinto1
+
+        Read bytes into a pre-allocated, writable bytes-like object b. Return the number of bytes read.
+        """
+        self._raise_if_closed()
+        self._raise_if_not_read()
+        if self._is_local_file:
+            if self._mode == "r":
+                return self._read_into_buffer(b)
+            size = self._file_stream.readinto1(b)
+            self._pos += size
+            return size
+        raise NotImplementedError(_NON_LOCAL_PATH_ERR_MSG)
+
+    def readline(self, size: int = -1) -> Sequence:
+        """
+        From https://docs.python.org/3/library/io.html#io.IOBase.readline
+
+        Read and return one line from the stream. If size is specified, at most size bytes will be read.
+        """
+        self._raise_if_closed()
+        self._raise_if_not_read()
+        if self._is_local_file:
+            content = self._file_stream.raw.readline(size)
+            self._pos += len(content)
+            return content
+        raise NotImplementedError(_NON_LOCAL_PATH_ERR_MSG)
+
+    def readlines(self, hint: int = -1) -> list[Sequence]:
+        """
+        From https://docs.python.org/3/library/io.html#io.IOBase.readlines
+
+        Read and return a list of lines from the stream. hint can be specified to control the number of lines read: no more
+        lines will be read if the total size (in bytes/characters) of all lines so far exceeds hint.
+
+        hint values of 0 or less, as well as None, are treated as no hint.
+
+        Note that it’s already possible to iterate on file objects using for line in file: ... without calling file.readlines().
+        """
+        self._raise_if_closed()
+        self._raise_if_not_read()
+        if self._is_local_file:
+            content = self._file_stream.raw.readlines(hint)
+            self._pos += sum(len(line) for line in content)
+            return content
+        raise NotImplementedError(_NON_LOCAL_PATH_ERR_MSG)
+
+    def seek(self, offset: int, whence: int = SEEK_SET) -> int:
         """
         See https://docs.python.org/3/library/io.html#io.IOBase.seek
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
 
-    def seekable(self) -> None:
+        Move the stream position to a new location given an offset and a starting position. SEEK_SET/0
+        indicates a position relative to the start of the file. SEEK_CUR/1 indicates a position relative
+        to the current stream position. SEEK_END/2 indicates a position relative to the end of the file.
+        Only supported in read mode.
+
+        Returns the new stream position. Not supported in write mode.
+        """
+        self._raise_if_closed()
+        self._raise_if_not_read()
+        if self._is_local_file:
+            if whence == SEEK_SET:
+                self._pos = offset
+            elif whence == SEEK_CUR:
+                self._pos = self._file_stream.tell() + offset
+            elif whence == SEEK_END:
+                self._pos = os.path.getsize(self._file_location) + offset
+            else:
+                raise NotImplementedError(f"Unsupported whence value {whence}")
+            if self._pos < 0:
+                raise ValueError(f"Negative seek position {self._pos}")
+            return self._file_stream.seek(self._pos, SEEK_SET)
+        raise NotImplementedError(_NON_LOCAL_PATH_ERR_MSG)
+
+    def seekable(self) -> bool:
         """
         See https://docs.python.org/3/library/io.html#io.IOBase.seekable
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
 
-    def tell(self) -> None:
+        Returns whether or not the stream is seekable.
+        """
+        self._raise_if_closed()
+        if self._is_local_file:
+            return self._mode in READ_MODES
+        raise NotImplementedError(_NON_LOCAL_PATH_ERR_MSG)
+
+    def tell(self) -> int:
         """
         See https://docs.python.org/3/library/io.html#io.IOBase.tell
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
 
-    def truncate(self, size: int | None = None) -> None:
+        Gets the current stream position. Not supported in write mode.
+        """
+        self._raise_if_closed()
+        self._raise_if_not_read()
+        if self._is_local_file:
+            return self._pos
+        raise NotImplementedError(_NON_LOCAL_PATH_ERR_MSG)
+
+    def truncate(self, size: int | None = None) -> int:
         """
         Not yet supported in UDF and Stored Procedures.
         """
+        self._raise_if_closed()
+        self._raise_if_not_write()
         raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
 
-    def write(self, b: bytes | bytearray | array.array) -> None:
+    def write(self, b: bytes | bytearray | array.array) -> int:
         """
         See https://docs.python.org/3/library/io.html#io.RawIOBase.write
-        """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
 
-    def writable(self) -> None:
+        Write the given bytes-like object, b, to the underlying raw stream, and return the number of bytes-like objects
+        written (e.g. unicode characters provided to a text input count as 1). The number of bytes should equal the input
+        bytes, since all bytes are written directly to the stream. Local testing support is not implemented yet.
+        """
+        raise NotImplementedError(_WRITE_MODE_ERR_MSG)
+
+    def writable(self) -> bool:
         """
         See https://docs.python.org/3/library/io.html#io.IOBase.writable
+
+        Returns whether or not the stream is writable.
         """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        self._raise_if_closed()
+        return self._file_stream.writable()
 
     def writelines(self, lines: Iterable[str] | list[str]) -> None:
         """
-        See https://docs.python.org/3/library/io.html#io.IOBase.writelines
+        From https://docs.python.org/3/library/io.html#io.IOBase.writelines
+
+        Write a list of lines to the stream. Line separators are not added, so it is usual for each of the lines provided to
+        have a line separator at the end. Local testing support is not implemented yet.
         """
-        raise NotImplementedError(_DEFER_IMPLEMENTATION_ERR_MSG)
+        raise NotImplementedError(_WRITE_MODE_ERR_MSG)